Step 4

+++ app/src/main/java/hexlet/code/controller/api/UserController.java
@@ -6,9 +6,11 @@ import hexlet.code.dto.UserUpdateDTO;
 import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.mapper.UserMapper;
 import hexlet.code.repository.UserRepository;
+import hexlet.code.utils.UserUtils;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -22,6 +24,9 @@ public class UserController {
 
     @Autowired
     private UserMapper userMapper;
+
+    @Autowired
+    private UserUtils userUtils;
 
     @GetMapping(path = "")
     public List<UserDTO> index() {
@@ -45,22 +50,31 @@ public class UserController {
     public UserDTO create(@Valid @RequestBody UserCreateDTO userData) {
         var user = userMapper.map(userData);
         userRepository.save(user);
-        var userDto = userMapper.map(user);
-        return userDto;
+        return userMapper.map(user);
     }
 
     @PutMapping(path = "/{id}")
     public UserDTO update(@Valid @RequestBody UserUpdateDTO userData, @PathVariable Long id) {
-        var user =  userRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("User with id " + id + " not found"));
-        userMapper.update(userData, user);
-        userRepository.save(user);
-        return userMapper.map(user);
+        var currentUser = userUtils.getCurrentUser();
+        if (currentUser.getId() == id) {
+            var user =  userRepository.findById(id)
+                    .orElseThrow(() -> new ResourceNotFoundException("User with id " + id + " not found"));
+            userMapper.update(userData, user);
+            userRepository.save(user);
+            return userMapper.map(user);
+        } else {
+            throw new AccessDeniedException("Access denied");
+        }
     }
 
     @DeleteMapping(path = "/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void delete(@PathVariable Long id) {
-        userRepository.deleteById(id);
+        var currentUser = userUtils.getCurrentUser();
+        if (currentUser.getId() == id) {
+            userRepository.deleteById(id);
+        } else {
+            throw new AccessDeniedException("Access denied");
+        }
     }
 }
